Refactor schema downloader tests for public vs. internal access
