+++ Tests/ApolloCodegenTests/ApolloSchemaPublicTests.swift
@@ -1,9 +1,9 @@
 import XCTest
 import ApolloTestSupport
 import ApolloCodegenTestSupport
-@testable import ApolloCodegenLib
+import ApolloCodegenLib
 
-class ApolloSchemaTests: XCTestCase {
+class ApolloSchemaPublicTests: XCTestCase {
   private var defaultOutputURL: URL {
     return CodegenTestHelper.outputFolderURL()
       .appendingPathComponent("schema.graphqls")
@@ -50,30 +50,6 @@ class ApolloSchemaTests: XCTestCase {
 
     let expectedOutputURL = sourceRoot.appendingPathComponent("\(schemaFileName).graphqls")
     XCTAssertEqual(configuration.outputURL, expectedOutputURL)
-  }
-
-  func testFormatConversion_givenIntrospectionJSON_shouldOutputValidSDL() throws {
-    let bundle = Bundle(for: type(of: self))
-    guard let jsonURL = bundle.url(forResource: "introspection_response", withExtension: "json") else {
-      throw XCTFailure("Missing resource file!", file: #file, line: #line)
-    }
-
-    try FileManager.default.apollo.createFolderIfNeeded(at: CodegenTestHelper.outputFolderURL())
-    let configuration = ApolloSchemaDownloadConfiguration(using: .introspection(endpointURL: TestURL.mockPort8080.url),
-                                                          outputFolderURL: CodegenTestHelper.outputFolderURL())
-
-    try ApolloSchemaDownloader.convertFromIntrospectionJSONToSDLFile(jsonFileURL: jsonURL, configuration: configuration)
-    XCTAssertTrue(FileManager.default.apollo.fileExists(at: configuration.outputURL))
-
-    let frontend = try ApolloCodegenFrontend()
-    let source = try frontend.makeSource(from: configuration.outputURL)
-    let schema = try frontend.loadSchemaFromSDL(source)
-
-    let authorType = try schema.getType(named: "Author")
-    XCTAssertEqual(authorType?.name, "Author")
-
-    let postType = try schema.getType(named: "Post")
-    XCTAssertEqual(postType?.name, "Post")
   }
 }
